service: Remove secret
This removes the secret used for client calls and puts the temporary trust store
certificate into place for pre-init request scenarios.

In addition this ensures that some of the methods like RemoteClusterMembers
can be invoked both during pre-init and after joining the cluster in case
a custom remote certificate is provided.
After pre-init is required for adding services to MicroCloud.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>