cmd/microcloud: Remove unused flag from preseed
This flag was outright ignored in MicroCloud, and there is not much
bneefit to adding support, so instead this removes the flag.

Reusing existing clusters is rather finicky and makes sense in
interactive setup because the user can configure it how they want, but
with preseed it doesn't make much sense, as the user would have had to
interact with the clusters already to know what fields to set in the
preseed.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>